refactor(gui): split ipc handler helpers
